pyproject.toml
[tool.poetry]
name = "education-agent"
version = "0.1.0"
description = "Minimal teaching repository for the OpenAI Agents SDK."
authors = ["fctechlabs"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai-agents = "*"
python-dotenv = "*"
requests = "*"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
